feat(persistence): se implementan operaciones CRUD para la entidad Película y agregar nuevas entidades
- Endpoints del controlador para crear, leer, actualizar y eliminar películas.
- Clases DTO y DAO para la transferencia de datos y la persistencia.
- Clases de repositorio y servicio para la interacción con la base de datos y la lógica de negocio.
- Clases adaptadoras para la integración con sistemas externos.

Además, este commit agrega nuevas entidades para Card, Order, Projection, Snack, Ticket y User. Estas entidades proporcionan las estructuras de datos necesarias para gestionar varios aspectos de la aplicación, como el procesamiento de pagos, la gestión de pedidos, la programación, las concesiones, la venta de entradas y las cuentas de usuario. Los detalles de la implementación de estas nuevas entidades se proporcionarán en commits posteriores.

+++ repository/src/main/java/pe/edu/certus/persistence/movie/port/jpa/ForManagingMovie.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import pe.edu.certus.persistence.movie.dao.MovieDAO;
 
 @Repository
-public interface ForManagingMovie extends JpaRepository<MovieDAO, Long> {
+public interface ForManagingMovie extends JpaRepository<MovieDAO, Integer> {
 }
